[pepsi-025] separated Main class to each method type
